+++ src/main/scala/com/goyeau/kubernetes/client/crd/JSON.scala
@@ -1,4 +1,4 @@
-package com.goyeau.kubernetes.client
+package com.goyeau.kubernetes.client.crd
 
 import io.circe.syntax._
 import io.circe.{Decoder, Encoder}
