+++ src/main/scala/com/goyeau/kubernetes/client/crd/RawExtension.scala
@@ -1,4 +1,4 @@
-package com.goyeau.kubernetes.client
+package com.goyeau.kubernetes.client.crd
 
 import io.circe.syntax._
